First batch of UI updates
The focus was on neumorphic buttons that work for both light and dark as well as state of the button (pressed)

Device specific enhancements

Colors and texts updated. Device specific paddings, margins and sizes enhanced

Dark appearance

Tweaks for the dark version of the onboarding UI

Update of the comments

Button UI updates

Colors updated, pressed button simulated and present in the DesignGuide.swift

Optimised Button's Layout

Onboarding footer was using the same settings for 3 buttons.