Issue #258 - User can take the backup test successfully more than once
Closes #258

this checks that when the user taps the finished button on the phrase displayed it has definitely not passed the test before going to the recovery flow.

Note: this should actually go to the next or previous screen according to the context.

That should be taken into consideration when implementing the profile screen that takes the user to the phrase diplay screen from that context.

Add //TODO comment with the permanent fix for the problem