[#611] Disable Send ZEC button when sync in progress
This disables the send button when the app is syncing. but if the user
is already there it won't change the underlying store to avoid unwanted
re-renders by SwiftUI engine.

Test reflect this situation. Also fixed a problem where the tests would
not reflect the correct state from the dependency injection.

Closes #611